+++ apps/fluent-tester/src/FluentTester/theme/CustomThemes.ts
@@ -1,10 +1,19 @@
+import { createAppleTheme } from '@fluentui-react-native/apple-theme';
 import { createDefaultTheme, ThemeOptions } from '@fluentui-react-native/default-theme';
 import { ThemeReference } from '@fluentui-react-native/theme';
+import { Platform } from 'react-native';
 import { applyBrand, OfficeBrand } from './applyBrand';
 import { applyTheme, ThemeNames } from './applyTheme';
+import { createAndroidTheme } from '@fluentui-react-native/android-theme';
 
 const themeOptions: ThemeOptions = { paletteName: 'TaskPane', appearance: 'dynamic' };
-const baseTheme = createDefaultTheme(themeOptions);
+
+const baseTheme = Platform.select({
+  android: createAndroidTheme(themeOptions),
+  ios: createAppleTheme(),
+  macos: createAppleTheme(),
+  default: createDefaultTheme(themeOptions),
+});
 
 export const lightnessOptions = [
   { label: 'Auto', value: 'dynamic' },
@@ -13,7 +22,7 @@ export const lightnessOptions = [
 ];
 
 export class TesterThemeReference extends ThemeReference {
-  private _themeName: ThemeNames = 'Fluent';
+  private _themeName: ThemeNames = 'Default';
   private _brand: OfficeBrand = 'Default';
 
   private options: ThemeOptions;