+++ packages/experimental/Icon/src/Icon.tsx
@@ -1,10 +1,9 @@
 import * as React from 'react';
 import { IconProps, SvgIconProps, FontIconProps } from './Icon.types';
-import { Image, ImageStyle } from 'react-native';
+import { Image, ImageStyle, Platform } from 'react-native';
 import { Text } from '@fluentui-react-native/text';
 import { SvgUri } from 'react-native-svg';
 import { mergeStyles } from '@fluentui-react-native/framework';
-import * as assetRegistry from 'react-native/Libraries/Image/AssetRegistry';
 import { stagedComponent, mergeProps, getMemoCache } from '@fluentui-react-native/framework';
 import { useTheme } from '@fluentui-react-native/theme-types';
 
@@ -18,8 +17,13 @@ function renderRasterImage(iconProps: IconProps) {
 }
 
 function fontFamilyFromFontSrcFile(fontSrcFile: string, fontFamily: string): string {
-  const asset = assetRegistry.getAssetByID(fontSrcFile);
-  return `${fontFamily}#${asset.httpServerLocation}/${asset.name}.${asset.type}`;
+  if (Platform.OS == 'windows') {
+    // This `${family}#${path}` notation is specific to WPF
+    const asset = Image.resolveAssetSource(+fontSrcFile);
+    return `${fontFamily}#${asset.uri}`;
+  } else {
+    return fontFamily;
+  }
 }
 
 const fontStyleMemoCache = getMemoCache();