+++ packages/theming/apple-theme/macos/MSFAppleThemeModule.m
@@ -1,0 +1,7 @@
+#import <React/RCTBridgeModule.h>
+
+@interface RCT_EXTERN_MODULE(MSFAppleThemeModule, NSObject)
+
+RCT_EXTERN_METHOD(getApplePartialThemeWithCallback:(RCTResponseSenderBlock)callback)
+
+@end