+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,118 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Sat, 12 Dec 2020 04:27:11 GMT and should not be manually modified.
+This log was last generated on Mon, 15 Mar 2021 06:53:17 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.21.0
+
+Mon, 15 Mar 2021 06:53:17 GMT
+
+### Minor changes
+
+- initial android theme (tamasane@gmail.com)
+
+## 0.20.0
+
+Fri, 12 Mar 2021 01:25:13 GMT
+
+### Minor changes
+
+- Get FURN Icon working on iOS (adam.gleitman@gmail.com)
+
+## 0.19.0
+
+Thu, 11 Mar 2021 19:48:16 GMT
+
+### Minor changes
+
+- Revamp ThemePickers in Header (saadnajmi2@gmail.com)
+
+## 0.18.1
+
+Wed, 10 Mar 2021 04:33:26 GMT
+
+### Patches
+
+- Add a native theming module for iOS (saadnajmi2@gmail.com)
+
+## 0.18.0
+
+Mon, 08 Mar 2021 21:55:00 GMT
+
+### Minor changes
+
+- initial two pager layout (tamasane@gmail.com)
+
+## 0.17.0
+
+Sat, 06 Mar 2021 00:20:05 GMT
+
+### Minor changes
+
+- Base theme on macOS changed to apple theme (saadnajmi2@gmail.com)
+- Fixed a lot of layout bugs (saadnajmi2@gmail.com)
+- Update to react-native 0.63 (30809111+acoates-ms@users.noreply.github.com)
+
+### Patches
+
+- Add Experimental Button to iOS tests (saadnajmi2@gmail.com)
+
+## 0.16.3
+
+Mon, 01 Mar 2021 22:53:12 GMT
+
+### Patches
+
+- Update deps (warleu@microsoft.com)
+
+## 0.16.2
+
+Mon, 01 Mar 2021 17:28:32 GMT
+
+### Patches
+
+- Remove metro logic from FURN, replacing it with calls to the @rnx-kit CLI. Use rnx-bundle for bundling, and rnx-start for running a bundle server. (afoxman@microsoft.com)
+
+## 0.16.0
+
+Fri, 19 Feb 2021 18:36:13 GMT
+
+### Minor changes
+
+- add tooltip support for native button (67026167+chiuam@users.noreply.github.com)
+
+## 0.15.0
+
+Fri, 05 Feb 2021 08:02:13 GMT
+
+### Minor changes
+
+- Added an Apple Theme (saadnajmi2@gmail.com)
+
+## 0.14.0
+
+Wed, 13 Jan 2021 20:04:26 GMT
+
+### Minor changes
+
+- mac native button (67026167+chiuam@users.noreply.github.com)
+
+## 0.13.12
+
+Wed, 13 Jan 2021 16:44:06 GMT
+
+### Patches
+
+- Added macOS implementation of experimental avatar (saadnajmi2@gmail.com)
+
+## 0.13.11
+
+Wed, 06 Jan 2021 18:41:50 GMT
+
+### Patches
+
+- Bump react-native-win32 (krsiler@microsoft.com)
 
 ## 0.13.9