+++ packages/components/RadioGroup/CHANGELOG.md
@@ -1,8 +1,44 @@
 # Change Log - @fluentui-react-native/radio-group
 
-This log was last generated on Fri, 30 Oct 2020 18:32:53 GMT and should not be manually modified.
+This log was last generated on Sat, 06 Mar 2021 00:20:05 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.0
+
+Sat, 06 Mar 2021 00:20:05 GMT
+
+### Minor changes
+
+- Update to react-native 0.63 (30809111+acoates-ms@users.noreply.github.com)
+
+### Patches
+
+- Base theme on macOS changed to apple theme (saadnajmi2@gmail.com)
+
+## 0.8.5
+
+Mon, 01 Mar 2021 22:53:12 GMT
+
+### Patches
+
+- Update deps (warleu@microsoft.com)
+
+## 0.8.2
+
+Wed, 06 Jan 2021 18:41:50 GMT
+
+### Patches
+
+- Bump react-native-win32 (krsiler@microsoft.com)
+
+## 0.8.1
+
+Wed, 04 Nov 2020 17:44:39 GMT
+
+### Patches
+
+- Selection now changes on a click as well as focus (safreibe@microsoft.com)
 
 ## 0.8.0