+++ packages/experimental/Button/CHANGELOG.md
@@ -1,8 +1,45 @@
 # Change Log - @fluentui-react-native/experimental-button
 
-This log was last generated on Fri, 23 Oct 2020 22:27:37 GMT and should not be manually modified.
+This log was last generated on Sat, 06 Mar 2021 00:20:05 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.0
+
+Sat, 06 Mar 2021 00:20:05 GMT
+
+### Minor changes
+
+- Update to react-native 0.63 (30809111+acoates-ms@users.noreply.github.com)
+
+### Patches
+
+- Base theme on macOS changed to apple theme (saadnajmi2@gmail.com)
+- Use ColorValue from @fluentui-react-native/theme-types (saadnajmi2@gmail.com)
+
+## 0.4.2
+
+Mon, 01 Mar 2021 22:53:12 GMT
+
+### Patches
+
+- Update deps (warleu@microsoft.com)
+
+## 0.4.0
+
+Wed, 13 Jan 2021 20:04:26 GMT
+
+### Minor changes
+
+- fix depcheck error by adding component-cache dependency for experimental button (67026167+chiuam@users.noreply.github.com)
+
+## 0.3.3
+
+Wed, 06 Jan 2021 18:41:50 GMT
+
+### Patches
+
+- Bump react-native-win32 (krsiler@microsoft.com)
 
 ## 0.3.0