+++ packages/experimental/Icon/CHANGELOG.md
@@ -1,8 +1,53 @@
 # Change Log - @fluentui-react-native/icon
 
-This log was last generated on Fri, 23 Oct 2020 22:27:37 GMT and should not be manually modified.
+This log was last generated on Fri, 12 Mar 2021 01:25:13 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.4.0
+
+Fri, 12 Mar 2021 01:25:13 GMT
+
+### Minor changes
+
+- Get FURN Icon working on iOS (adam.gleitman@gmail.com)
+
+## 0.3.0
+
+Sat, 06 Mar 2021 00:20:05 GMT
+
+### Minor changes
+
+- Update to react-native 0.63 (30809111+acoates-ms@users.noreply.github.com)
+
+### Patches
+
+- Base theme on macOS changed to apple theme (saadnajmi2@gmail.com)
+- Use ColorValue from @fluentui-react-native/theme-types (saadnajmi2@gmail.com)
+
+## 0.2.7
+
+Mon, 01 Mar 2021 22:53:12 GMT
+
+### Patches
+
+- Update deps (warleu@microsoft.com)
+
+## 0.2.4
+
+Wed, 06 Jan 2021 18:41:50 GMT
+
+### Patches
+
+- Bump react-native-win32 (krsiler@microsoft.com)
+
+## 0.2.3
+
+Wed, 04 Nov 2020 18:56:53 GMT
+
+### Patches
+
+- Fix import statement (kinhln@microsoft.com)
 
 ## 0.2.0